updated all components. Icons are not working
